+++ src/parser.js
@@ -7,6 +7,34 @@ function createError(message, line, column) {
   return error;
 }
 
+function parse(text, opts) {
+  let parseFunction;
+
+  if (opts.parser === "flow") {
+    parseFunction = parseWithFlow;
+  } else if (opts.parser === "typescript") {
+    parseFunction = parseWithTypeScript;
+  } else {
+    parseFunction = parseWithBabylon;
+  }
+
+  try {
+    return parseFunction(text);
+  } catch (error) {
+    const loc = error.loc;
+
+    if (loc) {
+      const codeFrame = require("babel-code-frame");
+      error.codeFrame = codeFrame(text, loc.line, loc.column + 1, {
+        highlightCode: true
+      });
+      error.message += "\n" + error.codeFrame;
+    }
+
+    throw error;
+  }
+}
+
 function parseWithFlow(text) {
   // Inline the require to avoid loading all the JS if we don't use it
   const flowParser = require("flow-parser");
@@ -97,4 +125,4 @@ function isProbablyJsx(text) {
   ].join(""), "m").test(text);
 }
 
-module.exports = { parseWithFlow, parseWithBabylon, parseWithTypeScript };
+module.exports = { parse };
